Add files via upload
Guess who finally figured out a minor error in the Bandits mod? One of the mid-range bandits was referencing a survivor lever shotgun, which was removed sometime between 0.D and 0.E. Crude fix, but I just appended it to the end of the monster file. Use it or not, as you see fit, but it does work and doesn't seem to break anything.