Fix copy-from target of swappable storage battery
The whole point of the swappable storage battery is being able to change them without tools. Without this fix the swappable battery needs a hacksaw to remove, like the regular storage battery. I'm honestly not sure if this entry is needed anymore, but at least it's not a pita anymore.